Moved more code to .cpp
