+++ dbms/src/Parsers/ASTCreateQuery.cpp
@@ -66,6 +66,19 @@ void ASTStorage::formatImpl(const FormatSettings & s, FormatState & state, Forma
 }
 
 
+class ASTColumnsElement : public IAST
+{
+public:
+    String prefix;
+    IAST * elem;
+
+    String getID(char c) const override { return "ASTColumnsElement for " + elem->getID(c); }
+
+    ASTPtr clone() const override;
+
+    void formatImpl(const FormatSettings & s, FormatState & state, FormatStateStacked frame) const override;
+};
+
 ASTPtr ASTColumnsElement::clone() const
 {
     auto res = std::make_shared<ASTColumnsElement>();
@@ -117,6 +130,7 @@ void ASTColumns::formatImpl(const FormatSettings & s, FormatState & state, Forma
     ASTExpressionList list;
 
     if (columns)
+    {
         for (const auto & column : columns->children)
         {
             auto elem = std::make_shared<ASTColumnsElement>();
@@ -124,7 +138,9 @@ void ASTColumns::formatImpl(const FormatSettings & s, FormatState & state, Forma
             elem->set(elem->elem, column->clone());
             list.children.push_back(elem);
         }
+    }
     if (indices)
+    {
         for (const auto & index : indices->children)
         {
             auto elem = std::make_shared<ASTColumnsElement>();
@@ -132,6 +148,7 @@ void ASTColumns::formatImpl(const FormatSettings & s, FormatState & state, Forma
             elem->set(elem->elem, index->clone());
             list.children.push_back(elem);
         }
+    }
 
     if (!list.children.empty())
         list.formatImpl(s, state, frame);
